backend/upload-file/requirements.txt
psycopg2-binary==2.9.9
python-multipart==0.0.6
boto3==1.34.34